+++ src/config.ts
@@ -1,249 +1,854 @@
-export const UNIVERSAL_DATA = {
-  rotation_degrees: 0,
-  zoom: 1,
-  paper_texture: "grey",
-  is_paper_shadow: true,
-  x_offset: 0,
-  y_offset: 0,
+import { nanoid } from "nanoid";
+type iDataInputRange = {
+  name: string;
+  type: "range";
+  value: number;
+  min: number;
+  max: number;
+  step: number;
+  suffix: string;
 };
 
-export const PAPER_TYPES = {
-  NEWSPAPER: {
-    name: "Newspaper (Spread Top)",
-    data: {
-      ...UNIVERSAL_DATA,
-      page_width_percentage: 80,
-      ink_color: "ink-black",
+type iDataInputPlainText = {
+  name: string;
+  type: "input";
+  value: string;
+  placeholder?: string;
+};
 
-      title: "THE LOREM IPSUM",
-      title_font: "font-sans",
-      title_size: "34",
-      line_height: "1",
-      top_margin: "0",
-      bottom_margin: "0",
+type iDataInputInkColorPicker = {
+  name: string;
+  type: "ink_color_picker";
+  value: string;
+};
 
-      banner_text_1: "Lorem ipsum",
-      banner_text_2: "Lorem ipsum",
-      banner_text_3: "Lorem ipsum",
-      banner_font: "font-serif",
-      banner_size: "16",
-      hide_top_banner_border: false,
-      hide_bottom_banner_border: false,
+type iDataInputFontPicker = {
+  name: string;
+  type: "font_picker";
+  value: string;
+};
 
-      headline: "MORBI TEMPUS MOLLIS IPSUM AT!",
-      headline_font: "font-serif",
-      headline_font_size: "34",
-      headline_line_height: "1",
-      headline_top_margin: "0",
-      headline_bottom_margin: "0",
+type iDataInputBoolean = {
+  name: string;
+  type: "boolean";
+  value: boolean;
+};
 
-      quote:
-        "Maecenas molestie ac, erat sed ultrices. Ut in vehicula est, ut malesuada eros! Nunc condimentum, aliquet ante nec venenatis.",
+type iDataInputPaperTexture = {
+  name: string;
+  type: "paper_texture";
+  value: string;
+};
 
-      quote_font: "font-serif",
-      quote_size: "32",
+type iDataInputImageFilter = {
+  name: string;
+  type: "image_filter";
+  value: string;
+};
 
-      main_copy: `Sed nec nisl eu massa accumsan facilisis ut a lacus. Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. Aenean at tristique nunc, nec ornare nibh. ![alt text](https://static.standard.co.uk/s3fs-public/thumbnails/image/2016/05/18/17/black_chronicles_exhibition_at_the_national_portrait_gallery.00_01_37_07.still001.jpg?width=1200) Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Suspendisse nec mauris aliquam, suscipit eros eget, imperdiet sapien. Sed bibendum molestie faucibus. Aenean faucibus placerat felis, nec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. ### Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. Duis tristique metus arcu, ac porttitor risus maximus at. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed nec nisl eu massa accumsan facilisis ut a lacus. Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. Aenean at tristique nunc, nec ornare nibh. Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Suspendisse nec mauris aliquam, suscipit eros eget, imperdiet sapien. Sed bibendum molestie faucibus. # Aenean faucibus placerat felis, nec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. Duis tristique metus arcu, ac porttitor risus maximus at. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed nec nisl eu massa accumsan facilisis ut a lacus. Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. ![alt text](http://images.summitmedia-digital.com/esquiremagph/images/2020/07/09/bearded-batman-1800s-MAIN.jpg) Duis tristique metus arcu, ac porttitor risus maximus at. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed nec nisl eu massa accumsan facilisis ut a lacus. Aenean at tristique nunc, nec ornare nibh. Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Sed euismod metus sed mauris accumsan. ### Aenean faucibus placerat felis Nnec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. Duis tristique metus arcu, ac porttitor risus maximus at. ![alt text](https://static.stacker.com/s3fs-public/00007074_0.png) Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed nec nisl eu massa accumsan facilisis ut a lacus. Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. Aenean at tristique nunc, nec ornare nibh. Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Suspendisse nec mauris aliquam, suscipit eros eget, imperdiet sapien. Sed bibendum molestie faucibus. Aenean faucibus placerat felis, nec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. ![alt text](https://media.npr.org/assets/img/2015/09/15/oldsmobile-loc_custom-d4496485e41ba2476a7392c5252f3d0a5cdce3f9.jpg) Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. Aenean at tristique nunc, nec ornare nibh. Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Suspendisse nec mauris aliquam, suscipit eros eget, imperdiet sapien. Sed bibendum molestie faucibus. Aenean faucibus placerat felis, nec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. Duis tristique metus arcu, ac porttitor risus maximus at. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos.`,
-      main_copy_columns: 3,
-      image_filter: "img-filter-grayscale",
-      is_main_copy_blurry: true,
-    },
-  },
-  NEWSPAPER_ALT: {
-    name: "Newspaper (Grid)",
-    data: {
-      ...UNIVERSAL_DATA,
-      gridData: {
-        totalColumns: 3,
-        gridGap: 10,
-        elements: [],
+type iDataInputTextArea = {
+  name: string;
+  type: "textarea";
+  value: string;
+  isMarkdown: boolean;
+};
+
+type iDataInputTextAlign = {
+  name: string;
+  type: "text_align";
+  value: string;
+};
+type iDataInputFontWeightPicker = {
+  name: string;
+  type: "font_weight_picker";
+  value: string;
+};
+type iDataInputSelect = {
+  name: string;
+  type: "select";
+  value: string;
+  options: any[];
+};
+type iDataInputRawImageUrl = {
+  name: string;
+  type: "raw_image_url";
+  value: string;
+};
+type iDataInputBlendMode = {
+  name: string;
+  type: "blend_mode";
+  value: string;
+};
+type iDataInputColor = {
+  name: string;
+  type: "color";
+  value: string;
+};
+
+export type Edge = "top" | "right" | "bottom" | "left";
+
+export type tHandoutData =
+  | iDataInputRange
+  | iDataInputPlainText
+  | iDataInputInkColorPicker
+  | iDataInputFontPicker
+  | iDataInputBoolean
+  | iDataInputPaperTexture
+  | iDataInputImageFilter
+  | iDataInputTextArea
+  | iDataInputTextAlign
+  | iDataInputFontWeightPicker
+  | iDataInputSelect
+  | iDataInputRawImageUrl
+  | iDataInputBlendMode
+  | iDataInputColor;
+
+export type iHandoutDefinition = {
+  name?: string;
+  caption?: string;
+  data: {
+    [key: string]:
+      | tHandoutData
+      | tHandoutData[]
+      | { [key: string]: tHandoutData }
+      | { [key: string]: tHandoutData }[];
+  };
+};
+
+export const isHandoutData = (obj: any): obj is tHandoutData => {
+  return obj.type !== undefined;
+};
+
+const fontSizeRange = (
+  name: string = "Font Size",
+  value: number = 34,
+  min: number = 26,
+  max: number = 200,
+  step: number = 1
+): iDataInputRange => ({
+  name,
+  type: "range",
+  value,
+  min,
+  max,
+  step: 1,
+  suffix: "px",
+});
+
+const fontPicker = (name: string = "Font"): iDataInputFontPicker => ({
+  name,
+  type: "font_picker",
+  value: "font-serif",
+});
+
+const _rotationDegrees: iDataInputRange = {
+  name: "Rotation",
+  type: "range",
+  value: 0,
+  min: -60,
+  max: 60,
+  step: 1,
+  suffix: "°",
+};
+const _zoom: iDataInputRange = {
+  name: "Zoom",
+  type: "range",
+  value: 1,
+  min: 0.1,
+  max: 6,
+  step: 0.05,
+  suffix: "",
+};
+const _paperTexture: iDataInputPaperTexture = {
+  name: "Paper Texture",
+  type: "paper_texture",
+  value: "grey",
+};
+const _isPaperShadow: iDataInputBoolean = {
+  name: "Enable paper inset shadow?",
+  type: "boolean",
+  value: true,
+};
+const _xOffset: iDataInputRange = {
+  name: "X Offset",
+  type: "range",
+  value: 0,
+  min: -100,
+  max: 100,
+  step: 1,
+  suffix: "%",
+};
+const _yOffset: iDataInputRange = {
+  name: "Y Offset",
+  type: "range",
+  value: 0,
+  min: -100,
+  max: 100,
+  step: 1,
+  suffix: "%",
+};
+const textAreaWithMarkdown = (
+  name: string = "Text Area"
+): iDataInputTextArea => {
+  return {
+    name,
+    type: "textarea",
+    value: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+
+## Pellentesque nisl ipsum, sodales at velit sit amet, tempor sagittis orci.`,
+
+    isMarkdown: true,
+  };
+};
+export const POSITIONING_DATA = {
+  rotation_degrees: _rotationDegrees,
+  zoom: _zoom,
+  x_offset: _xOffset,
+  y_offset: _yOffset,
+};
+
+export const NEWSPAPER = {
+  name: "Newspaper (Front Page)",
+  caption:
+    "The front page of a newspaper. Configurable title, banners, headline, call-out and copy. Make the copy blurry for a more dramatic effect.",
+  data: {
+    positioning: {
+      ...POSITIONING_DATA,
+      y_offset: {
+        ..._yOffset,
+        value: 0,
       },
-      page_width_percentage: 80,
-      ink_color: "ink-black",
+    },
 
-      title: "THE LOREM IPSUM",
-      title_font: "font-sans",
-      title_size: "34",
+    page_width_percentage: {
+      name: "Page Width",
+      type: "range",
+      value: 80,
+      min: 0,
+      max: 300,
+      step: 1,
+      suffix: "%",
+    },
+    paper_texture: _paperTexture,
+    paper_tint: {
+      name: "Paper Colour Tint",
+      type: "color",
+      value: "#FFFFFF",
+    },
 
-      banner_text_1: "Lorem ipsum",
-      banner_text_2: "Lorem ipsum",
-      banner_text_3: "Lorem ipsum",
-      banner_font: "font-serif",
-      banner_size: "16",
-      hide_top_banner_border: false,
-      hide_bottom_banner_border: false,
+    ink_color: {
+      name: "Ink Colour",
+      type: "ink_color_picker",
+      value: "ink-black",
+    },
+    is_paper_shadow: _isPaperShadow,
+    title: {
+      title: {
+        name: "Title",
+        value: "THE LOREM IPSUM",
+        type: "input",
+      },
+      title_font: {
+        name: "Font",
+        value: "font-sans",
+        type: "font_picker",
+      },
+      title_font_size: fontSizeRange("Font Size", 50),
+      line_height: {
+        name: "Line Height",
+        type: "range",
+        value: 1,
+        min: -5,
+        max: 5,
+        step: 0.1,
+        suffix: "em",
+      },
+      top_margin: {
+        name: "Top Margin",
+        type: "range",
+        value: 0,
+        min: -100,
+        max: 100,
+        step: 1,
+        suffix: "px",
+      },
+      bottom_margin: {
+        name: "Bottom Margin",
+        type: "range",
+        value: 0,
+        min: -100,
+        max: 100,
+        step: 1,
+        suffix: "px",
+      },
+    },
+    banner_texts: {
+      banner_text_1: {
+        name: "Banner Text 1",
+        type: "input",
+        value: "Integer a egestas",
+      },
+      banner_text_2: {
+        name: "Banner Text 1",
+        type: "input",
+        value: "Duis sodales",
+      },
+      banner_text_3: {
+        name: "Banner Text 1",
+        type: "input",
+        value: "Quisque imperdiet",
+      },
+      banner_font: fontPicker(),
+      banner_size: fontSizeRange("Banner Text Size", 16),
+      hide_top_banner_border: {
+        name: "Hide Top Banner Border",
+        type: "boolean",
+        value: false,
+      },
+      hide_bottom_banner_border: {
+        name: "Hide Bottom Banner Border",
+        type: "boolean",
+        value: false,
+      },
+    },
 
-      headline: "MORBI TEMPUS MOLLIS IPSUM AT!",
-      headline_font: "font-serif",
+    headline: {
+      headline: {
+        name: "Headline",
+        type: "input",
+        value: "ALIQUAM EROS AUGUE, COMMODO EGET RHONCUS NEC!",
+      },
+      headline_font: {
+        name: "Font",
+        value: "font-serif",
+        type: "font_picker",
+      },
+      headline_font_size: fontSizeRange("Font Size", 34),
+      headline_line_height: {
+        name: "Line Height",
+        type: "range",
+        value: 80,
+        min: 0,
+        max: 100,
+        step: 1,
+        suffix: "em",
+      },
+      headline_top_margin: {
+        name: "Top Margin",
+        type: "range",
+        value: 80,
+        min: 0,
+        max: 100,
+        step: 1,
+        suffix: "em",
+      },
+      headline_bottom_margin: {
+        name: "Bottom Margin",
+        type: "range",
+        value: 80,
+        min: 0,
+        max: 100,
+        step: 1,
+        suffix: "em",
+      },
+    },
 
-      quote:
-        "Maecenas molestie ac, erat sed ultrices. Ut in vehicula est, ut malesuada eros! Nunc condimentum, aliquet ante nec venenatis.",
+    quote: {
+      quote: {
+        name: "Quote / Call Out",
+        type: "input",
+        value:
+          '"Maecenas molestie ac, erat sed ultrices. Ut in vehicula est, ut malesuada eros! Nunc condimentum, aliquet ante nec venenatis"',
+      },
+      quote_font: {
+        name: "Font",
+        value: "font-serif",
+        type: "font_picker",
+      },
+      quote_font_size: fontSizeRange("Font Size", 20),
+    },
 
-      quote_font: "font-serif",
-      quote_size: "32",
-
-      main_copy: `Sed nec nisl eu massa accumsan facilisis ut a lacus. Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. Aenean at tristique nunc, nec ornare nibh. ![alt text](https://static.standard.co.uk/s3fs-public/thumbnails/image/2016/05/18/17/black_chronicles_exhibition_at_the_national_portrait_gallery.00_01_37_07.still001.jpg?width=1200) Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Suspendisse nec mauris aliquam, suscipit eros eget, imperdiet sapien. Sed bibendum molestie faucibus. Aenean faucibus placerat felis, nec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. ### Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. Duis tristique metus arcu, ac porttitor risus maximus at. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed nec nisl eu massa accumsan facilisis ut a lacus. Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. Aenean at tristique nunc, nec ornare nibh. Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Suspendisse nec mauris aliquam, suscipit eros eget, imperdiet sapien. Sed bibendum molestie faucibus. # Aenean faucibus placerat felis, nec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. Duis tristique metus arcu, ac porttitor risus maximus at. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed nec nisl eu massa accumsan facilisis ut a lacus. Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. ![alt text](http://images.summitmedia-digital.com/esquiremagph/images/2020/07/09/bearded-batman-1800s-MAIN.jpg) Duis tristique metus arcu, ac porttitor risus maximus at. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed nec nisl eu massa accumsan facilisis ut a lacus. Aenean at tristique nunc, nec ornare nibh. Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Sed euismod metus sed mauris accumsan. ### Aenean faucibus placerat felis Nnec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. Duis tristique metus arcu, ac porttitor risus maximus at. ![alt text](https://static.stacker.com/s3fs-public/00007074_0.png) Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Sed nec nisl eu massa accumsan facilisis ut a lacus. Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. Aenean at tristique nunc, nec ornare nibh. Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Suspendisse nec mauris aliquam, suscipit eros eget, imperdiet sapien. Sed bibendum molestie faucibus. Aenean faucibus placerat felis, nec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. ![alt text](https://media.npr.org/assets/img/2015/09/15/oldsmobile-loc_custom-d4496485e41ba2476a7392c5252f3d0a5cdce3f9.jpg) Nam velit lorem, porta convallis dignissim ac, porta non quam. Quisque tincidunt nulla eu mi ornare, in vestibulum magna consequat. Sed euismod metus sed mauris accumsan, in pretium arcu euismod. Aenean at tristique nunc, nec ornare nibh. Nulla commodo suscipit tortor. Morbi eu urna feugiat, bibendum libero commodo, pharetra nulla. Suspendisse nec mauris aliquam, suscipit eros eget, imperdiet sapien. Sed bibendum molestie faucibus. Aenean faucibus placerat felis, nec vulputate velit faucibus at. Proin eget malesuada eros, et posuere leo. Nam non gravida lectus. Maecenas commodo, ex sed rhoncus dignissim, mauris purus mollis diam, nec tempus ex ligula sed diam. Duis tristique metus arcu, ac porttitor risus maximus at. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos.`,
-      main_copy_columns: 3,
-      image_filter: "img-filter-grayscale",
-      is_main_copy_blurry: true,
+    main_copy: {
+      main_copy_content: textAreaWithMarkdown("Main Copy"),
+      main_copy_columns: {
+        name: "Main Copy Columns",
+        type: "range",
+        value: 3,
+        min: 1,
+        max: 4,
+        step: 1,
+        suffix: "",
+      },
+      image_filter: {
+        name: "Image Effect",
+        type: "image_filter",
+        value: "none",
+      },
+      is_main_copy_blurry: {
+        name: "Is Main Copy Blurry?",
+        type: "boolean",
+        value: true,
+      },
+      text_align: {
+        name: "Text Align",
+        type: "text_align",
+        value: "text-justify",
+      },
     },
   },
-  NEWSPAPER_CLIPPING: {
-    name: "Newspaper Clipping",
-    data: {
-      ...UNIVERSAL_DATA,
-      page_width: 400,
-      page_height: 700,
-      font: "font-serif",
-      ink_color: "ink-black",
-      image_filter: "none",
-      prefix_copy:
-        "Nam sed tincidunt velit. Ut ac ultricies magna. Sed gravida metus est, eget bibendum purus pharetra sed. Maecenas scelerisque mauris ut ligula vestibulum auctor. Nulla pharetra arcu eros, eu fermentum tortor aliquet ac. In ut arcu accumsan, dictum massa id, pretium dui.",
-      main_copy: `---
+} as const;
+export const NEWSPAPER_CLIPPING = {
+  name: "Newspaper Clipping (Vertical)",
+  caption:
+    "A simplified version of a newspaper 'clipping' - a classic vertical slice.",
+  data: {
+    positioning: {
+      ...POSITIONING_DATA,
+      y_offset: {
+        ..._yOffset,
+        value: 0,
+      },
+    },
+    dimensions: {
+      page_width: {
+        name: "Page Width",
+        type: "range",
+        value: 400,
+        min: 100,
+        max: 1800,
+        step: 1,
+        suffix: "px",
+      },
+      page_height: {
+        name: "Page Height",
+        type: "range",
+        value: 700,
+        min: 100,
+        max: 1200,
+        step: 1,
+        suffix: "px",
+      },
+    },
+    paper: {
+      paper_texture: _paperTexture,
+      paper_tint: {
+        name: "Paper Colour Tint",
+        type: "color",
+        value: "#FFFFFF",
+      },
+      is_paper_shadow: _isPaperShadow,
+    },
+    prefix_copy: textAreaWithMarkdown("Prefix Copy"),
+    is_prefix_blurry: {
+      name: "Is Prefix Blurry?",
+      type: "boolean",
+      value: true,
+    },
+    main_copy: textAreaWithMarkdown("Main Copy"),
+    suffix_copy: textAreaWithMarkdown("Suffix Copy"),
+    is_suffix_blurry: {
+      name: "Is Suffix Blurry?",
+      type: "boolean",
+      value: true,
+    },
 
-### Aenean ac semper!
-
-Nulla nisi felis, ultrices nec orci at, sollicitudin viverra mauris. 
-
-Mauris in pharetra nunc. Nullam semper metus mi, a iaculis nisi maximus sed. Proin sed turpis ipsum.
-
----`,
-      suffix_copy: `## LOREM IPSUM!
-
-Nulla nisi felis, ultrices nec orci at, sollicitudin viverra mauris. 
-
-Mauris in pharetra nunc. Nullam semper metus mi, a iaculis nisi maximus sed. Proin sed turpis ipsum.`,
-
-      is_prefix_blurry: true,
-      is_suffix_blurry: true,
+    font: fontPicker(),
+    ink_color: {
+      name: "Ink Colour",
+      type: "ink_color_picker",
+      value: "ink-black",
+    },
+    image_filter: {
+      name: "Image Effect",
+      type: "image_filter",
+      value: "none",
     },
   },
-  WANTED_POSTER: {
-    name: "Wanted Poster",
-    data: {
-      ...UNIVERSAL_DATA,
-      page_width: 500,
-      paper_texture: "beige",
-      image_url:
-        "https://images.unsplash.com/photo-1596473785107-fef2debc419c?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=965&q=80",
-      image_filter: "none",
-      headline: "WANTED!",
-      headline_font: "font-serif",
-      subtitle: "Etiam tempor, massa blandit cursus efficitur",
-      subtitle_font: "font-serif",
-      subtitle2: "DEAD OR ALIVE",
-      subtitle2_font: "font-sans",
+} as const;
+export const CHARACTER_CARD = {
+  name: "Character Card",
+  caption: "An easy-to-use character card. A picture with 3 lines of text.",
+  data: {
+    positioning: {
+      ...POSITIONING_DATA,
+      zoom: {
+        ..._zoom,
+        value: 1,
+      },
     },
-  },
-  HANDWRITTEN_LETTER: {
-    name: "Handwritter Letter",
-    data: {
-      ...UNIVERSAL_DATA,
-      page_width_percentage: 80,
-      padding: "p-8",
-      font_size: "prose-lg",
-      ink_color: "ink-black",
-      font_weight: "font-normal",
-      text_alignment: "text-left",
-      main_copy: `Dear Lorem Ipsum,
-
-Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-
-Nam at dignissim enim. Nam at ipsum diam. Sed bibendum nibh mi, at ultricies augue eleifend eget. In vel eros vitae ex elementum iaculis. Phasellus fermentum, elit viverra vulputate euismod, ipsum massa sagittis arcu, et placerat felis purus ut ex. Curabitur venenatis et velit sed semper. Morbi vitae rutrum tortor, eget varius metus. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
-
-Nulla suscipit sapien at ligula suscipit ultricies. Pellentesque sit amet vulputate sapien, et aliquet augue. 
-
-_Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas._
-
-Nulla scelerisque, sem vel efficitur gravida, nunc velit sagittis sapien, egestas interdum nunc lectus id velit. Donec sem nisl, maximus at orci quis, auctor faucibus elit. Aliquam consectetur rutrum augue eget tincidunt.
-
-Sincerely, Lorem Ipsum`,
-      font: "font-indie-flower",
+    page_width: {
+      name: "Page Width",
+      type: "range",
+      value: 700,
+      min: 100,
+      max: 1500,
+      step: 1,
+      suffix: "px",
     },
-  },
-  TICKET: {
-    name: "Ticket",
-    data: {
-      ...UNIVERSAL_DATA,
-      // page_width_percentage: 80,
-      page_width: 500,
-      page_height: 280,
-      sawtooth_border: "sawtooth-border-none",
-      rounded_corners: "rounded-none",
-
-      hide_left_margin_copy: false,
-      left_margin_copy: "CARTER AND DARK LTD.",
-
-      hide_top_content: false,
-      top_copy: "Nulla suscipit sapien",
-
-      hide_right_margin_copy: false,
-      right_margin_copy: "Ligula suscipit ultricies",
-
-      hide_middle_left_copy: false,
-      middle_left_copy: "Nulla suscipit sapien",
-
-      hide_middle_right_copy: false,
-      middle_right_copy: "Nulla suscipit sapien",
-
-      hide_botom_copy: false,
-      botom_copy: "Nulla suscipit sapien",
-
-      // font_size: "prose-lg",
-      // prefix: "To my Dearest,",
-      // prefix_alignment: "left",
-
-      font_weight: "font-normal",
-      font: "font-serif",
+    paper_texture: {
+      ..._paperTexture,
+      value: "beige-3",
     },
-  },
-  BLANK_PAGES: {
-    name: "Blank Pages",
-    data: {
-      ...UNIVERSAL_DATA,
-    },
-  },
-  BOOK_COVER: {
-    name: "Book Cover",
-    data: {
-      ...UNIVERSAL_DATA,
-      book_cover_template: "/images/ragged-journal-cover.png",
-      book_cover_template_image_effect: "none",
-      text_effect: "none",
-      font: "font-im-fell-display",
-      zoom_level_percentage: "30",
-      prose_size: "prose",
-      ink_color: "ink-red",
-      extra_left_margin: 0,
-      main_copy: `# MYSTICAL ARTIFACTS
-
-  ## A STUDY IN THE STRANGE AND WEIRD
-
-  ### A. COBALT ET.AL`,
-    },
-  },
-  NPC_CARD: {
-    name: "Character Card",
-    data: {
-      ...UNIVERSAL_DATA,
-      page_width: 300,
-      paper_texture: "beige",
-      image_url:
+    is_paper_shadow: _isPaperShadow,
+    image_url: {
+      name: "Image URL",
+      type: "input",
+      value:
         "https://i.pinimg.com/564x/49/c1/4d/49c14d528399386e820dd116a25590b2.jpg",
-      image_filter: "none",
-      image_border: "none",
-      ink_color: "ink-black",
-      text_line_one: "Lorem Ipsum",
-      text_line_two: "Lorem Ipsum",
-      text_line_three: "Lorem Ipsum",
-      font: "font-serif",
-      font_weight: "font-normal",
-      rounded_corners: "rounded-none",
-      text_alignment: "text-left",
+    },
+    ink_color: {
+      name: "Ink Colour",
+      type: "ink_color_picker",
+      value: "ink-black",
+    },
+    image_filter: {
+      name: "Image Effect",
+      type: "image_filter",
+      value: "none",
+    },
+    text_line_one: {
+      name: "Line 1",
+      type: "input",
+      value: "Lorem Ipsum",
+    },
+    text_line_two: {
+      name: "Line 2",
+      type: "input",
+      value: "Lorem Ipsum",
+    },
+    text_line_three: {
+      name: "Line 3",
+      type: "input",
+      value: "Lorem Ipsum",
+    },
+    font: {
+      name: "Font",
+      type: "font_picker",
+      value: "font-serif",
+    },
+    font_size: fontSizeRange("Font Size (Relative)", 24),
+    font_weight: {
+      name: "Font Weight",
+      type: "font_weight_picker",
+      value: "font-normal",
+    },
+    text_align: {
+      name: "Text Align",
+      type: "text_align",
+      value: "text-left",
     },
   },
-  // PUB_SIGN: {
-  //   name: "Pub Sign",
-  //   data: {},
-  // },
+} as const;
+export const PLAIN_LETTER = {
+  name: "Plain Letter",
+  caption:
+    "A simple plain sheet of paper. Completely configurable via markdown.",
+  data: {
+    positioning: {
+      ...POSITIONING_DATA,
+      zoom: {
+        ..._zoom,
+        value: 1,
+      },
+    },
+    page_width: {
+      name: "Page Width",
+      type: "range",
+      value: 850,
+      min: 100,
+      max: 1500,
+      step: 1,
+      suffix: "px",
+    },
+    is_paper_shadow: _isPaperShadow,
+    paper_texture: _paperTexture,
+    paper_tint: {
+      name: "Paper Colour Tint",
+      type: "color",
+      value: "#FFFFFF",
+    },
+    ink_color: {
+      name: "Ink Colour",
+      type: "ink_color_picker",
+      value: "ink-black",
+    },
+    padding: {
+      name: "Padding",
+      type: "range",
+      min: 0,
+      max: 100,
+      value: 50,
+      step: 1,
+      suffix: "px",
+    },
+    paragraph: [
+      {
+        id: nanoid(),
+        main_copy: {
+          name: "Main Copy",
+          type: "textarea",
+          value: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
+
+## Pellentesque nisl ipsum, sodales at velit sit amet, tempor sagittis orci. 
+
+Nullam cursus congue magna, pulvinar commodo massa ornare quis. Etiam eleifend fermentum mauris at aliquet. Aliquam ac augue nunc. 
+
+Mauris mollis bibendum erat, et ultrices tortor pellentesque vitae. 
+
+# Suspendisse sed accumsan augue.
+
+Mauris orci tortor, semper nec purus ac, rhoncus mollis massa. Cras euismod dignissim libero ut luctus. Ut mattis ut tellus quis aliquet. In hac habitasse platea dictumst.          
+`,
+          isMarkdown: true,
+        },
+        font: {
+          name: "Font",
+          type: "font_picker",
+          value: "font-serif",
+        },
+        font_size: fontSizeRange("Font Size (Relative)", 12),
+        font_weight: {
+          name: "Font Weight",
+          type: "font_weight_picker",
+          value: "font-normal",
+        },
+        text_align: {
+          name: "Text Align",
+          type: "text_align",
+          value: "text-left",
+        },
+      },
+    ],
+    paragraph_gap: {
+      name: "Paragraph Gap",
+      type: "range",
+      min: 0,
+      value: 0,
+      max: 100,
+      step: 1,
+      suffix: "px",
+    },
+  },
+} as const;
+export const BOOK_COVER = {
+  name: "Book Cover",
+  caption: "A book cover, with markdown-configurable cover text",
+  data: {
+    book_cover_template: {
+      name: "Book Cover Template",
+      type: "select",
+      value: "/images/book-cover-1.avif",
+      options: [
+        { label: "Book Cover 1", value: "/images/book-cover-1.avif" },
+        { label: "Book Cover 2", value: "/images/book-cover-2.avif" },
+        { label: "Book Cover 3", value: "/images/book-cover-3.png" },
+        { label: "Book Cover 4", value: "/images/book-cover-4.png" },
+      ],
+    },
+    positioning: {
+      ...POSITIONING_DATA,
+      zoom: {
+        ..._zoom,
+        value: 1,
+      },
+    },
+    ink_color: {
+      name: "Ink Colour",
+      type: "ink_color_picker",
+      value: "ink-white",
+    },
+    font: {
+      name: "Font",
+      type: "font_picker",
+      value: "font-serif",
+    },
+    font_size: fontSizeRange("Font Size (Relative)", 30),
+    font_weight: {
+      name: "Font Weight",
+      type: "font_weight_picker",
+      value: "font-normal",
+    },
+    text_left_margin: {
+      name: "Text Left Margin",
+      type: "range",
+      value: 13,
+      min: -100,
+      max: 100,
+      step: 1,
+      suffix: "px",
+    },
+    text_align: {
+      name: "Text Align",
+      type: "text_align",
+      value: "text-center",
+    },
+    text_effect: {
+      name: "Text Effect",
+      type: "blend_mode",
+      value: "blend-mode-normal",
+    },
+    main_copy: {
+      name: "Main Copy",
+      type: "textarea",
+      isMarkdown: true,
+      value: `# MYSTICAL ARTIFACTS
+
+## A STUDY IN THE STRANGE AND WEIRD
+
+### A. COBALT ET.AL`,
+    },
+  },
+} as const;
+
+export const LABELLED_LIQUID = {
+  name: "Labelled Liquid",
+  caption: "A bottle, vial, or other labelled liquid",
+  data: {
+    image_template: {
+      name: "Image Template",
+      type: "select",
+      value: "/images/labelled-container-1.avif",
+      options: [
+        {
+          label: "Round Bottle w/ Stopper",
+          value: "/images/labelled-container-1.avif",
+        },
+        { label: "Potion 2", value: "/images/labelled-container-2.avif" },
+        {
+          label: "Jar w/ Green Stuff",
+          value: "/images/labelled-container-3.avif",
+        },
+        { label: "Blue Bottle", value: "/images/labelled-container-4.avif" },
+        {
+          label: "Blue Bottle Alt",
+          value: "/images/labelled-container-5.avif",
+        },
+        { label: "Potion 6", value: "/images/labelled-container-6.avif" },
+      ],
+    },
+    image_hue_filter: {
+      name: "Image Hue Filter (Rotate)",
+      type: "range",
+      value: 0,
+      min: 0,
+      max: 360,
+      step: 1,
+      suffix: "°",
+    },
+    image_width: {
+      name: "Image Width",
+      type: "range",
+      value: 400,
+      min: 100,
+      max: 1000,
+      step: 2,
+      suffix: "px",
+    },
+    image_rotation: {
+      name: "Image Rotation",
+      type: "range",
+      value: 0,
+      min: -360,
+      max: 360,
+      step: 1,
+      suffix: "°",
+    },
+    positioning: {
+      ...POSITIONING_DATA,
+      zoom: {
+        ..._zoom,
+        value: 1,
+      },
+    },
+    ink_color: {
+      name: "Ink Colour",
+      type: "ink_color_picker",
+      value: "ink-white",
+    },
+    font: {
+      name: "Font",
+      type: "font_picker",
+      value: "font-serif",
+    },
+    font_size: fontSizeRange("Font Size (Relative)", 12, 1, 100, 0.1),
+    font_weight: {
+      name: "Font Weight",
+      type: "font_weight_picker",
+      value: "font-normal",
+    },
+    text_top_margin: {
+      name: "Text Top Margin",
+      type: "range",
+      value: 60,
+      min: -100,
+      max: 150,
+      step: 1,
+      suffix: "%",
+    },
+    text_left_margin: {
+      name: "Text Left Margin",
+      type: "range",
+      value: 13,
+      min: -100,
+      max: 100,
+      step: 1,
+      suffix: "px",
+    },
+    text_width: {
+      name: "Text Width",
+      type: "range",
+      value: 80,
+      min: -100,
+      max: 150,
+      step: 1,
+      suffix: "%",
+    },
+    text_rotation: {
+      name: "Text Rotation",
+      type: "range",
+      value: 0,
+      min: -360,
+      max: 360,
+      step: 1,
+      suffix: "°",
+    },
+    text_align: {
+      name: "Text Align",
+      type: "text_align",
+      value: "text-center",
+    },
+    text_effect: {
+      name: "Text Effect",
+      type: "blend_mode",
+      value: "blend-mode-normal",
+    },
+    main_copy: {
+      name: "Main Copy",
+      type: "textarea",
+      isMarkdown: true,
+      value: `# MYSTICAL ARTIFACTS
+
+## A STUDY IN THE STRANGE AND WEIRD
+
+### A. COBALT ET.AL`,
+    },
+  },
+} as const;
+
+export enum eHandoutDefinitions {
+  NEWSPAPER = "NEWSPAPER",
+  NEWSPAPER_CLIPPING = "NEWSPAPER_CLIPPING",
+  CHARACTER_CARD = "CHARACTER_CARD",
+  PLAIN_LETTER = "PLAIN_LETTER",
+  BOOK_COVER = "BOOK_COVER",
+  LABELLED_LIQUID = "LABELLED_LIQUID",
+}
+
+export const ALL_HANDOUT_DEFINITIONS = {
+  NEWSPAPER,
+  NEWSPAPER_CLIPPING,
+  CHARACTER_CARD,
+  PLAIN_LETTER,
+  BOOK_COVER,
+  LABELLED_LIQUID: LABELLED_LIQUID,
 };
